requirements.txt
pydantic==2.5.3
pytest==7.4.4
pymongo==4.6.1
python-dotenv==1.0.0